Fix kube-update GO version detection (#377)
GO version was not updated on the latest Kube patch update because
the local repo was not on the expected GO version for the prior Kube version.
Update the logic so that current GO version is determined by what is
actually being used by the repo and set the current GO version of the repo
to what it should be using for the latest Kube release.

Also added logic to prevent downgrading GO to an earlier version of GO. 
Some of the older branches in this repo are on a newer version of GO than the 
community due to various dependency requirements.  Do not downgrade GO back  
to the community GO version when the community pushed out new update.